Improve ToTrace formatting
The Extension method ToTrace, for use on Exceptions, did not correctly indent messages with multiple lines. It also left an extra \r character when printing multiline traces.
This could lead to some unusual output.

This commit resolves the formatting work, and makes a few small changes to file paths and test project namespaces.